Made "go up/down/left/right" by default move cursor by 1 line/character. Also refactored the 2-worded commands back to original design.
