+++ src/commandProcessor/commandProcessor.ts
@@ -61,7 +61,6 @@ export const processCommand = (input: string) => {
     console.log("This is a navigation command");
 
     const secondWordIdx = 1;
-    const beyondSecondWordIdx = 2;
     const thirdWordIdx = 2;
     const beyondThirdWordIdx = 3;
     const fifthWordIdx = 4;
@@ -69,18 +68,9 @@ export const processCommand = (input: string) => {
 
     try {
       if (inputCmdArray[secondWordIdx] in NavigationKeyword) {
-        if (inputCmdArray.length === 2) {
-          // Yes, this is meant to be splice(secondWordIdx)
-          const remaining = inputCmdArray.splice(secondWordIdx)[0];
-          processNavigation(prefix, remaining);
-        } else {
-          const prefixNotCamel = inputCmdArray
-            .slice(0, beyondSecondWordIdx)
-            .join(" ");
-          const prefixCamel = camelize(prefixNotCamel);
-          const remaining = inputCmdArray.slice(beyondSecondWordIdx).join(" ");
-          processNavigation(prefixCamel, remaining);
-        }
+        // Yes, this is meant to be splice(secondWordIdx)
+        const remaining = inputCmdArray.splice(secondWordIdx).join(" ");
+        processNavigation(prefix, remaining);
       } else if (inputCmdArray[thirdWordIdx] in NavigationKeyword) {
         const prefixNotCamel = inputCmdArray
           .slice(0, beyondThirdWordIdx)
